% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chchpd_auth.R
\name{chchpd_check_rstudio_server}
\alias{chchpd_check_rstudio_server}
\title{Check whether RStudio is running in server mode:}
\usage{
chchpd_check_rstudio_server()
}
\description{
Check whether RStudio is running in server mode:
}
