% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_chchpd_cache}
\alias{clear_chchpd_cache}
\title{Clear cached data and force to download afresh from Googlesheets.}
\usage{
clear_chchpd_cache()
}
\description{
Clear cached data and force to download afresh from Googlesheets.
}
\examples{
\dontrun{
clear_chchpd_cache()
}
}
