% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{DEPRECATED_import_neuropsyc}
\alias{DEPRECATED_import_neuropsyc}
\title{Import (old) neuropsyc data.}
\usage{
DEPRECATED_import_neuropsyc(concise = TRUE)
}
\arguments{
\item{concise}{If \code{TRUE}, return only the selected variables (e.g. MoCA,
global z, domain z). If \code{FALSE}, also return all individual test
scores, allowing more detailed analyses.}
}
\value{
A dataframe containing neuropsyc scores.
}
\description{
\code{import_neuropsyc} Return various test results from the OLD neuropsyc
spreadsheet. This should no longer be used, as it refers only to an outdated
data source. The new \code{import_neuropsyc()} function imports data from a
sheet which is regularly exported from the REDCap database.
}
\examples{
\dontrun{
np = import_neuropsyc()
}
}
