% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{get_chchpd_options}
\alias{get_chchpd_options}
\title{View current `chchpd` package options}
\usage{
get_chchpd_options()
}
\value{
A list of options and their current values.
}
\description{
\code{get_chchpd_options} View a list of pacakge options, such as whether
data should be cached and for how long, and whether to suppress Google sheets
warnings.
}
\examples{
\dontrun{
get_chchpd_options()
}
}
