% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_bloods}
\alias{import_bloods}
\title{Import data from biosamples collection sessions.}
\usage{
import_bloods()
}
\value{
A dataframe containing the blood samples and other data.
}
\description{
\code{import_bloods}
}
\details{
This PET information is held in a fixed Google Sheet and so no parameters are
necessary for this function.
}
\examples{
\dontrun{
bloods <- import_bloods()
}
}
