% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_hallucinations}
\alias{import_hallucinations}
\title{Import data from the Sydney PsycH-Q hallucinations questionnaire.}
\usage{
import_hallucinations()
}
\value{
A dataframe containing the hallucination data.
}
\description{
\code{import_hallucinations} Get patient and significant-other ratings of
hallucinations and delusions, using the scale developed by Prof. Simon Lewis'
team in Sydney.
}
\examples{
\dontrun{
hallucinations <- import_hallucinations()
}
}
