% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_MDS_UPDRS}
\alias{import_MDS_UPDRS}
\title{Import MDS-UPDRS scores.}
\usage{
import_MDS_UPDRS(concise = TRUE)
}
\arguments{
\item{concise}{If \code{TRUE}, return only H & Y stage and totals for Parts
I, II, and III. If \code{FALSE}, retain all individual item scores.}
}
\value{
A dataframe containing the scores.
}
\description{
\code{import_MDS_UPDRS} Import the total scores calculated from the
MDS-UPDRS individual items for each part and the Hoehn & Yahr stage.
}
\examples{
\dontrun{
full_mds_updrs = import_MDS_UPDRS(concise = FALSE)
}
}
