% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_medications}
\alias{import_medications}
\title{Import medications data.}
\usage{
import_medications()
}
\arguments{
\item{concise}{If \code{TRUE}, return only the calculated LED (levodopa
equivalent daily dose). If \code{FALSE}, also return all individual
medication doses.}
}
\value{
A dataframe containing medication doses.
}
\description{
\code{import_medications} Return doses of anti-parkinsonian medications in
use at the time of the assessment session.
}
\examples{
\dontrun{
meds = import_medications()
}
}
