% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_motor_scores}
\alias{import_motor_scores}
\title{Import UPDRS/MDS-UPDRS motor scores.}
\usage{
import_motor_scores()
}
\value{
A dataframe containing the motor score information.
}
\description{
\code{import_motor_scores} Import the total Part III motor score and the Hoehn
& Yahr stage from the UPDRS/MDS-UPDRS datasets
}
\details{
At the beginnining of the study, the original 1987 UPDRS was still in use.
This function in turn calls the \code{import_old_UPDRS} function that will
import data from that assessment and scale it using an established method so
that the Part III total becomes comparable to the Part III MDS-UPDRS total
from later sessions. For scaling details, see Goetz et al. 2012. Mov
Disorders, 27(10):1239-1242

To bind the two datasets, we need to drop all but the summary scores. If you
want access to individual items scores, use the individual functions
\code{import_old_UPDRS} or \code{import_MDS_UPDRS}
}
\examples{
\dontrun{
updrs <- import_motor_scores()
}
}
