% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_neuropsyc}
\alias{import_neuropsyc}
\title{Import neuropsyc data.}
\usage{
import_neuropsyc(concise = TRUE)
}
\arguments{
\item{concise}{If \code{TRUE}, return only selected variables (e.g. MoCA,
global z, domain z). If \code{FALSE}, also return all individual test
scores, allowing more detailed analyses.}
}
\value{
A dataframe containing neuropsyc scores. Notes for particular variables:

\describe{
  \item{global_z}{Default global z, equal to global_z_no_language.}
  \item{global_z_no_language}{The mean of four domain mean scores, excluding 
     the language domain. Requires at least one score in each of the four 
     domains, otherwise returns NA. NA is also returned if the session is 
     marked incomplete in REDCap (data not fully collected or entered).}
  \item{global_z_with_language}{The mean of five domain mean scores. 
     Requires at least one score in each of the five domains, otherwise returns 
     NA. NA is also returned if the session is marked incomplete.}
  \item{global_z_grand_mean}{The grand mean of all z-scores across all domains. 
     Is returned as NA if there are fewer than four z-scores or if the session 
     is marked as incomplete.}
  \item{n_z_scores}{The number of z_scores available for calculating
     the global_z scores and cognitive status.}
  \item{mci_criteria_requirements_met}{Whether the number of test requirements
     by domain for Level-II criteria have been met when determining the 
     cognitive status.}
  \item{cognitive_status}{Normal-range cognition (N),  Mild cognitive 
     impairment (MCI) or dementia (D). Will return NA when there are
     fewer than four z-scores or the session is marked as incomplete
     and doesn't have an overriding dementia diagnosis. Several of these
     diagnoses don't strictly meet Level-II criteria and this can be 
     checked by looking at the mci_criteria_requirements_met variable.
     Depending on your analysis some of these may need to be excluded.
     In some cases this is due to missing language tests from when the 
     study started, in other cases it is due to a greater number of tests
     not being administered (and can be examined by looking at n_z_scores).}
}
}
\description{
\code{import_neuropsyc} Return various test results from the neuropsyc
data, as periodically exported from the REDCap database.
}
\examples{
\dontrun{
np <- import_neuropsyc()
}
}
