% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_old_UPDRS}
\alias{import_old_UPDRS}
\title{Import original UPDRS (1987 version) scores.}
\usage{
import_old_UPDRS()
}
\arguments{
\item{concise}{If \code{TRUE}, return only H & Y stage and totals for Parts
I, II, and III. If \code{FALSE}, retain all individual item scores.}
}
\value{
A dataframe containing the scores.
}
\description{
\code{import_old_UPDRS} Import original UPDRS scale data, which was used in
the earliest years of  data collection. The raw items here are different, and
so we also calculate Part II and Part III scores which have been scaled to
match the MDS-UPDRS equivalents using the method described by from Goetz et
al. 2012. Mov Disorders, 27(10):1239-1242.
}
\examples{
\dontrun{
part_III <- import_motor_scores()
}
}
