% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_participants}
\alias{import_participants}
\title{Import participant demographics}
\usage{
import_participants(anon = FALSE)
}
\arguments{
\item{anon}{If \code{TRUE}, return the anonymous ID. Otherwise, return the
in-house, non-secure ID.}
}
\value{
A dataframe containing the participant data.
}
\description{
\code{import_participants} Access the participant information that should be
constant across all sessions (for example, date of birth, sex, etc).
}
\details{
This information is exported periodically from the Alice database.
}
\examples{
\dontrun{
participants = import_participants()
}
}
