% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_PET}
\alias{import_PET}
\title{Import metadata on PET scans}
\usage{
import_PET(exclude = TRUE)
}
\value{
A dataframe containing the PET metadata.
}
\description{
\code{import_PET_scan_numbers} accesses the scan number and dose information
associated with each PET scan session.
}
\details{
This PET information is held in a fixed Google Sheet and so no parameters are
necessary for this function.
}
\examples{
\dontrun{
pet <- import_PET_scan_numbers()
}
}
