% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_sessions}
\alias{import_sessions}
\title{Import links between sessions and studies}
\usage{
import_sessions(
  from_study = NULL,
  exclude = TRUE,
  print_exclude_summary = TRUE
)
}
\arguments{
\item{from_study}{Optionally specify the name of a specific study to limit
the records returned to just those from that study. Inspect the sessions
spreadsheet to see the valid options. e.g. \code{'PET'} or
\code{'Follow-up'} for the Progression study. Can also specify a list to
include sessions from multiple studies, e.g.
\code{c('Follow-up', 'PD DNA')}.}

\item{exclude}{If \code{TRUE}, don't return sessions that have been excluded
from a given study. For safety, we default to \code{TRUE} so that you don't
get junk data unless you explicitly ask for it. You might set this to
\code{FALSE} in order to get all records if you need to document
exclusions, for example to create a participant recruitment and exclusion
flowchart.}

\item{print_exclude_summary}{If \code{TRUE}, print a table that lists
the number of records that are excluded from each study. If this looks
problematic, you could set \code{exclude = FALSE} to import all records to
identify which ones were excluded.}
}
\value{
A dataframe containing the session-group-study data.
}
\description{
\code{import_sessions} Access a table that links a given session to the
(possibly multiple) studies of which it is a part, and the group within that
study to which the associated participant belongs.
}
\details{
This information is exported periodically from the Alice database.
}
\examples{
\dontrun{
sessions = import_sessions()
}
}
