% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{tabulate_duplicates}
\alias{tabulate_duplicates}
\title{Tabulate number of duplicated values in a variable}
\usage{
tabulate_duplicates(df, varname, print_results = FALSE)
}
\arguments{
\item{df}{The dataframe containing the variable of interest.}

\item{varname}{The variable of interest. Must be given as a string to avoid
dplyr's non-standard evaluation.}

\item{print_results}{If \code{TRUE}, print the result as a
\code{knitr::kable} formatted table as well as returning the dataframe.}
}
\description{
\code{tabulate_duplicates} Produce a table of which values are duplicated
within a specified variable in a dataframe, and their number of repetitions.
}
\examples{
\dontrun{
duplicate_bloods = tabulate_duplicates(bloods, 'subject_id')
}
}
