% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{google_authenticate}
\alias{google_authenticate}
\title{Allow Google Drive authorisation via a server}
\usage{
google_authenticate(email = gargle::gargle_oauth_email(),
  use_server = FALSE)
}
\arguments{
\item{email}{Allows you to nominate a particular nzbri.org e-mail address to
use for authentication. This can save you having to specify it manually each
time the script is run, but shouldn't be used in reproducible workflows
(where other users have to be able to run the code).}

\item{use_server}{If \code{TRUE}, use the copy/paste authentication process.
If \code{FALSE}, use fully browser-contained authentication.

The ability to specify \code{use_server = FALSE} is provided for completeness
but shouldn't be necessary in most cases (maybe your script is set up to run
either locally or on the server, and this would allow the authentication
method to be specified conditionally).}
}
\value{
No return value: the function initiates a process which results in
the writing of a \code{.httr-oauth} token file to disk.
}
\description{
\code{google_authenticate} To access data on a Google drive requires that the
user is authenticated, to ensure that you are entitled to view it.
}
\details{
If using RStudio on a server, then you should run this function before
attempting to import any datasets from the Google drive. Specifying
`use_server = TRUE` will institute a different authentication process.
}
