% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_MRI}
\alias{import_MRI}
\title{Import metadata on MRI scans}
\usage{
import_MRI(exclude = TRUE)
}
\arguments{
\item{exclude}{If \code{TRUE}, don't return records marked as excluded.}
}
\value{
A dataframe containing the MRI metadata.
}
\description{
\code{import_MRI} accesses the scan number and associated information
 from each MRI scanning session.
}
\examples{
\dontrun{
mri = import_MRI()
}
}
