% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_neuropsyc}
\alias{import_neuropsyc}
\title{Import neuropsyc data.}
\usage{
import_neuropsyc(concise = TRUE)
}
\arguments{
\item{concise}{If \code{TRUE}, return only the selected variables (e.g. MoCA,
global z, domain z). If \code{FALSE}, also return all individual test
scores, allowing more detailed analyses.}
}
\value{
A dataframe containing medication doses.
}
\description{
\code{import_neuropsyc} Return various test results from the neuropsyc
spreadsheet.
}
\examples{
\dontrun{
np = import_neuropsyc()
}
}
