% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_neuropsyc}
\alias{import_neuropsyc}
\title{Import neuropsyc data.}
\usage{
import_neuropsyc(concise = TRUE)
}
\arguments{
\item{concise}{If \code{TRUE}, return only selected variables (e.g. MoCA,
global z, domain z). If \code{FALSE}, also return all individual test
scores, allowing more detailed analyses.}
}
\value{
A dataframe containing neuropsyc scores. Notes for particular variables:

\describe{
  \item{global_z}{A mean of all global z-scores. Is returned as NA if there
     are fewer than four z-scores or if the session is marked as 
     incomplete in REDCap (data not fully collected or entered).}
  \item{global_z_no_language}{Similar to global_z except that z-scores from
     the language domain are excluded and matches the definition in some 
     historical papers.}
  \item{n_z_scores}{The number of z_scores that have been used in calculating
     the global_z mean and cognitive status.}
  \item{mci_criteria_requirements_met}{Whether the number of test requirements
     by domain for Level-II criteria have been met when determining the 
     cognitive status.}
  \item{cognitive_status}{Normal-range cognition (N),  Mild cognitive 
     impairment (MCI) or dementia (D). Will return NA when there are
     fewer than four z-scores or the session is marked as incomplete
     and doesn't have an overriding dementia diagnosis. Several of these
     diagnoses don't strictly meet Level-II criteria and this can be 
     checked by looking at the mci_criteria_requirements_met variable.
     Depending on your analysis some of these may need to be excluded.
     In some cases this is due to missing language tests from when the 
     study started, in other cases it is due to a greater number of tests
     not being administered (and can be examined by looking at n_z_scores).}
}
}
\description{
\code{import_neuropsyc} Return various test results from the neuropsyc
data, as periodically exported from the REDCap database.
}
\examples{
\dontrun{
np <- import_neuropsyc()
}
}
