% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_participants}
\alias{import_participants}
\title{Import participant demographics}
\usage{
import_participants(anon_id = FALSE, identifiers = FALSE)
}
\arguments{
\item{anon_id}{If \code{TRUE}, return an anonymised \code{subject_id}.
Otherwise, return the in-house, non-secure ID.
Currently, this option is not useful if other datasets are being imported,
as they still contain the standard ID. We need to implement a global setting
that dictates whether all functions return anonymised IDs.}

\item{identifiers}{If \code{TRUE}, include identifying information (names,
dates of birth & death, etc). This should not be used routinely for research.
i.e. this information should only ever be needed for study management
purposes.
\code{anon_id} must be set to \code{FALSE} for this parameter to have any
effect.}
}
\value{
A dataframe containing the participant data.
}
\description{
\code{import_participants} Access the participant information that should be
constant across all sessions (for example, age at diagnosis, sex, etc).
}
\details{
This information is exported periodically from the Alice database.
}
\examples{
\dontrun{
participants <- import_participants()
}
}
