% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_participants}
\alias{import_participants}
\title{Import participant demographics}
\usage{
import_participants(anon_id = TRUE, identifiers = FALSE)
}
\arguments{
\item{anon_id}{If \code{TRUE}, return the anonymous ID. Otherwise, return the
in-house, non-secure ID.}

\item{identifiers}{If \code{TRUE}, include identifying information (names,
dates of birth & death, etc). This should not be used routinely for research.
i.e. this information should only ever be needed for study management
purposes.
\code{anon_id} must be set to \code{FALSE} for this parameter to have any
effect.}
}
\value{
A dataframe containing the participant data.
}
\description{
\code{import_participants} Access the participant information that should be
constant across all sessions (for example, date of birth, sex, etc).
}
\details{
This information is exported periodically from the Alice database.
}
\examples{
\dontrun{
participants = import_participants()
}
}
