% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_sessions}
\alias{import_sessions}
\title{Import links between sessions and studies}
\usage{
import_sessions()
}
\value{
A dataframe containing the session-group-study data.
}
\description{
\code{import_sessions} Access a table that links a given session to the
(possibly multiple) studies of which it is a part, and the group within that
study to which the associated participant belongs.
}
\details{
This information is exported priodically from the Alice database.
}
\examples{
\dontrun{
sessions = import_sessions()
}
}
